makefile flag on

+++ srcs/ResponseHandler.cpp
@@ -227,7 +227,7 @@ bool ResponseHandler::isCgi()
 
 	for (std::vector<std::string>::iterator it = ext.begin(); it != ext.end(); it++)
 	{
-		int index = uri.find(*it);
+		size_t index = uri.find(*it);
 		if (index != std::string::npos && uri.compare(index, it->length(), *it) == 0)
 		{
 			int queryIndex;
@@ -238,7 +238,7 @@ bool ResponseHandler::isCgi()
 			}
 			metaVariable["SCRIPT_NAME"] = uri.substr(0, index + it->length());
 			uri = uri.substr(index);
-			int pathIndex = uri.find('/');
+			size_t pathIndex = uri.find('/');
 			if (pathIndex != std::string::npos)
 			{
 				metaVariable["PATH_INFO"] = uri.substr(pathIndex);
